[IMP] web_editor: update color picker layout
In order to save space and gain clarity, the colorpicker elements are
now displayed in two separate tabs: "theme" and "custom".
It is still possible to display the combinations tab instead of the
theme one (for the background color option for example).

Part of https://github.com/odoo/odoo/pull/67545
task-2476601