[FIX] mass_mailing: fix display condition of mass.mailing alert ribbon
Oversight of b8724b8ad3d6db57380e52f57b4e22876b1afdb7

This commit fixes the display condition of the small alert block on top of the
mass.mailing form view.

The condition was not working because it would hide the alert if:
('failed', '=', 0)
OR
	('state', '!=', 'in_queue')
	AND
	('sent', '=', 0)
	AND
	('scheduled', '=', 0)

Instead, we want to hide it if:
('failed', '=', 0)
AND
('state', '!=', 'in_queue')
AND
('sent', '=', 0)
AND
('scheduled', '=', 0)

This would for example prevent showing that the mailing is "scheduled" or that
it has successfully been sent with various mailing statistics.

Task-2457227

closes odoo/odoo#67752

X-original-commit: 98786b70d92d8a3bdfb961945f475b64899cc2e0
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>