[IMP] crm_iap_lead: improve error handling for lead mining requests
This commit improves the crm.iap.lead.mining.request error handling by
displaying the encountered error (currently one of "no results" or
"insufficient credits") within a warning block on top of the form view.

This allows the user to quickly understand what's going on and change his
parameters to get results or buy additional credits if he's out.

Furthermore, this commit simplifies the technical implementation by avoiding to
re-create a separate crm.iap.lead.mining.request when we encounter an issue and
instead allow the user to modify the current request with an error state.

UPG PR odoo/upgrade#2261
Task-2478427

closes odoo/odoo#67574

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>